--- fs_impl_test.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-	"os"
-)
-
-func TestEmptyServer(t *testing.T) {
-	t.Parallel()
-	srv, dir, err := newTestServer()
-	if err != nil {
-		t.Errorf("new test server: %v", err)
-		return
-	}
-	defer os.RemoveAll(dir)
-
-	req := httptest.NewRequest(http.MethodGet, "/api/v1/fs/cat/1/2", nil)
-	w := httptest.NewRecorder()
-
-	srv.ServeHTTP(w, req)
-
-	res := w.Result()
-	defer res.Body.Close()
-	_, err = ioutil.ReadAll(res.Body)
-
-	if res.StatusCode != 400 {
-		t.Errorf("expected to fail. got %v", res.StatusCode)
-	}
-}
